fix(api): fix build compilation errors in server setup & resource logic
- Fix WebApplication.CreateBuilder syntax in
  Program.cs

- Add missing LastCollectedAt property to Mine
  model

- Explicitly cast decimal resource yields to long
  for material quantity updates

+++ MiningGame.API/Services/GameService.cs
@@ -132,7 +132,8 @@ public async Task CollectResourcesAsync(Guid playerId)
             var material = player.Materials.FirstOrDefault(m => m.Type.ToString() == mine.Type.ToString());
             if (material != null)
             {
-                material.Quantity += generatedAmount;
+                // CHQ: Gemini - variable labeled with long
+                material.Quantity += (long)generatedAmount;
             }
 
             mine.LastCollectedAt = now;
